Obtaining fapi status from sp properties
